+++ src/main/java/com/lims/manage/erp/service/EntrustService.java
@@ -78,6 +78,6 @@ public interface EntrustService {
      * @param entrustmentId
      * @return
      */
-    EntrustAddVo getEntrustHistoryDetail(Integer entrustmentId);
+    EntrustAddVo getEntrustHistoryDetail(Long entrustmentId);
 
 }
